+++ src/components/common/Table/TableComponent.module.css
@@ -6,7 +6,7 @@
   background-color: #ffffff;
 }
 
-.custom-table .ant-table-thead th {
+.customTable .ant-table-thead th {
   background-color: #e6f4ff !important;
   color: white;
 }
