+++ src/components/common/Table/TableComponent.tsx
@@ -23,7 +23,7 @@ export default function TableComponent<T>({
 
   return (
     <Table<T>
-      className="custom-table"
+      className={styles.customTable}
       bordered
       columns={columns}
       dataSource={datasource}
