+++ src/screens/employee/HistoryScreen.tsx
@@ -1,9 +1,9 @@
 import { message, TablePaginationConfig, Typography } from "antd";
 import React, { useEffect, useState } from "react";
-import { EmployeeTransactionHistoryColumnConfig } from "../../features/transactions/transactionType";
 import EmployeeDetailTransactionModal from "../../features/transactions/components/DetailTransaction/EmployeeDetailTransactionModal";
 import { useHistory } from "../../features/transactions/stores/useTransactionHistory";
 import TableComponent from "../../components/common/Table/TableComponent";
+import { EmployeeTransactionHistoryColumnConfig } from "../../features/transactions/components/TableTransactionColumnConfig/EmployeeTransactionColumnConfig";
 
 const { Title } = Typography;
